+++ fixtures/c/linked_list.c
@@ -1,7 +1,11 @@
 #include <stdlib.h>
 
 int main() {
-    int *head; int *temp; int *node;
+    // Cells must be 8 bytes so a single slot can hold either a value
+    // or a pointer-shaped next link. `long long` is 8 bytes on every
+    // target; plain `long` would only be 4 on Windows LLP64 and the
+    // next-pointer cast would truncate to 32 bits there.
+    long long *head; long long *temp; long long *node;
     int sum; int i;
 
     head = 0;
@@ -10,12 +14,12 @@ int main() {
     // Create a list of 5 nodes: [4, 3, 2, 1, 0]
     for (i = 0; i < 5; i++) {
         // One slot for value, one for the next pointer. Stored as a
-        // flat int[2]; the cast on the next-pointer write tells the
-        // type-checker we know we're stuffing a pointer-shaped value
-        // into a word-sized cell.
-        node = malloc(sizeof(int) + sizeof(int *));
-        node[0] = i;            // data
-        node[1] = (int)head;    // next
+        // flat long long[2]; the cast on the next-pointer write tells
+        // the type-checker we know we're stuffing a pointer-shaped
+        // value into a word-sized cell.
+        node = malloc(sizeof(long long) + sizeof(long long *));
+        node[0] = i;                // data
+        node[1] = (long long)head;  // next
         head = node;
     }
 
@@ -23,7 +27,7 @@ int main() {
     temp = head;
     while (temp != 0) {
         sum = sum + temp[0];
-        temp = (int *)temp[1];
+        temp = (long long *)temp[1];
     }
 
     return sum; // Expected: 4+3+2+1+0 = 10
