+++ fixtures/c/pthread_create.c
@@ -31,7 +31,11 @@ int main() {
     int *handle;
     int *create;
     int *join;
-    int tid;
+    // pthread_t is 8 bytes on every host we target; under  `int`
+    // is 4 bytes, so passing `&tid` to pthread_create would let the
+    // 8-byte write clobber the next stack slot. Use `long` for the
+    // handle storage.
+    long tid;
     int *retval;
 
     handle = dlopen(0, 2);              // RTLD_NOW
